+++ markwon-ext-tables/src/main/java/io/noties/markwon/ext/tables/TableSpan.java
@@ -1,0 +1,7 @@
+package io.noties.markwon.ext.tables;
+
+/**
+ * @since $nap;
+ */
+public class TableSpan {
+}